Flytter filtrer- og sammenlikn-operasjoner til nye hjem
